+++ Project/USERS/HardWare/HD7279.h
@@ -58,6 +58,6 @@ extern unsigned char  xdata realCode[];
 void HD7279SendByte(unsigned char sendData);
 unsigned char HD7279ReceiveByte(void);
 void HD7279Write(unsigned char cmd , unsigned char dataSend);
-void HD7279ShowInt(int showValue);
+void HD7279ShowInt(long showValue);
 void HD7279ShowFloat(double showValue);
 #endif
